WEAT-28 Covered network module by tests added json fake apis

+++ core/models/src/main/java/io/wookoo/models/weather/weekly/additional/WeeklyWeatherModel.kt
@@ -16,7 +16,7 @@ data class WeeklyWeatherModel(
     val dayLightDuration: List<Double>,
     val sunshineDuration: List<Double>,
     val uvIndexMax: List<Double>,
-    val precipitationProbabilityMax: List<Double>,
+    val precipitationProbabilityMax: List<Int>,
     val precipitationData: List<PrecipitationModel>,
     val windData: List<WindModel>,
 )
